refactor(ir): add SymbolAccess type to the symbol table
